beta version done

+++ gui_controller/live_cpt_data.py
@@ -1,3 +1,10 @@
+# this module implements a subclass of CPTdata which is particularly suited
+# for live stream. it is essentially the same as the CPTdata, except the
+# functions for processing data track which indices have thus far been
+# computed and only process indices corresponding to new data
+# whenever extract_candidates() is called. 
+
+
 import controller_config as config
 import cpt_tools
 
@@ -7,10 +14,15 @@ import time
 import os
 
 
-MCP_CAL_X = 1.29
-MCP_CAL_Y = 1.31
 
+# header_labels = np.concatenate( ( ['Z', 'A', 'timestamp', 'duration',
+#                                    'num_penning_ejects', 'num_mcp_hits' ],
+#                                   cpt_tools.TaborParams.flattened_header_labels(),
+#                                   [ 'header_key' ] ) )
+# header_labels = ', '.join( header_labels ) 
 
+ # self.Z, self.A, self.timestamp, self.duration,
+ #                                    self.num_penning_ejects, self.num_mcp_hits
 
 class LiveCPTdata( cpt_tools.CPTdata ) :
 
@@ -40,13 +52,18 @@ class LiveCPTdata( cpt_tools.CPTdata ) :
         self.num_mcp_hits = 0
         self.num_penning_ejects = 0
 
-        self.tdc.clear() 
+        self.num_rollovers = 0 
+        self.last_rollover = 0
+
+        self.tdc.clear()
+        # self.start_time = time.time() 
         
         # self.num_data = 0 
 
         
         # duration of session in seconds 
         self.duration = 0
+        self.date_str = datetime.now().strftime( '%Y-%m-%d_%H-%M-%S' ) 
 
         
     # @jit
@@ -59,12 +76,13 @@ class LiveCPTdata( cpt_tools.CPTdata ) :
     
         if config.BENCHMARK :
             start = time.time() 
-            
+
+        # rollovers would more appropriately be named "rollover_mask".
+        # for each hit of the TDC, this is 1 if the hit was a rollover,
+        # otherwise 0.
         rollovers = self.tdc.rollovers[ : self.tdc.num_data_in_buf ] 
 
         channel_indices = np.where( rollovers == 0 )[0]
-
-        # self.sort_data()
         
         cur_trig_idx = -1
 
@@ -76,7 +94,9 @@ class LiveCPTdata( cpt_tools.CPTdata ) :
         num_pos_channels_detected = 0 
 
         i = 0
-        while( i < len( channel_indices ) ) :         
+        
+        while( i < len( channel_indices ) ) :
+            
             idx = channel_indices[i]
 
             chan = self.tdc.channels[ idx ]
@@ -104,6 +124,8 @@ class LiveCPTdata( cpt_tools.CPTdata ) :
             # new mcp hit: found an event candidate 
             elif chan == 7 :
                 mcp_trigger_time = hit_time
+                mcp_trigger_timestamp = self.tdc.timestamps[ idx ] 
+                
                 tof = self.compute_tof( self.current_trig_time, mcp_trigger_time )
                 # self.all_tofs[ self.num_mcp_hits ] = tof 
                 self.num_mcp_hits += 1
@@ -129,24 +151,45 @@ class LiveCPTdata( cpt_tools.CPTdata ) :
                         if not pos_channel_buf[ chan ] : 
                             pos_channel_buf[ chan ] = hit_time
                             num_pos_channels_detected += 1 
+
+                    # catch unknown errors issued by the TDC. i have observed
+                    # that ignoring them doesn't cause any problems. 
                     except : 
-                        self.tdc.print_bin( self.tdc.data_buf[ idx ] )
-                        print( 'chan: ', chan )
-                        
+                        # self.tdc.print_bin( self.tdc.data_buf[ idx ] )
+                        # print( 'chan: ', chan )
+                        pass
 
                     # found enough data for a calculation 
                     if num_pos_channels_detected == 4 :
 
-                        cur_mcp_positions = self.compute_mcp_positions( pos_channel_buf )
-                                                
-                        self.mcp_positions[ self.num_events ] = cur_mcp_positions
+                        # cur_mcp_positions = self.compute_mcp_positions( pos_channel_buf )
+                        # sums = self.compute_sums( pos_channel_buf, mcp_trigger_time )
+                        
+                        self.delay_times[ self.num_events ] = self.compute_delay_times(
+                            pos_channel_buf, mcp_trigger_time )
+
+                        self.penning_eject_indices[ self.num_events ] = self.num_penning_ejects
+
+                        self.timestamps[ self.num_events ] = mcp_trigger_timestamp
+                        
+                        # self.mcp_positions[ self.num_events ] = cur_mcp_positions
                         self.tofs[ self.num_events ] = tof
+
+                        # self.sums[ self.num_events ] = self.compute_sums( pos_channel_buf,
+                        #                                                   mcp_trigger_time ) 
+
                         self.num_events += 1 
                         mcp_trigger_reached = 0 
 
             i += 1
 
+        # compute everything that can be computed in tandem from delay data
+        self.compute_mcp_positions()
+        self.compute_timestamp_differences() 
         self.compute_polar()
+        self.compute_sums()
+        self.compute_diff_xy() 
+        
         self.apply_cuts()
         self.num_cut_data_prev = self.num_cut_data
         self.num_events_prev = self.num_events
@@ -160,77 +203,22 @@ class LiveCPTdata( cpt_tools.CPTdata ) :
             
         self.tdc.num_data_in_buf = 0 
 
-    
-    
-    # @jit( nopython = 1 ) 
-    def compute_mcp_positions( self, pos_channel_buf ) :
-        # print( pos_channel_buf ) 
-        mcp_pos = np.zeros(2) 
-        mcp_pos[0] =  0.5 * MCP_CAL_X * 25 * ( pos_channel_buf[0] - pos_channel_buf[1] ) * 1e-3
-        mcp_pos[1] =  0.5 * MCP_CAL_Y * 25 * ( pos_channel_buf[2] - pos_channel_buf[3] ) * 1e-3
-        return mcp_pos
+
         
-
-    # def compute_sums( self, pos_channel_buf ) :
-    #     ret = np.zeros(2) 
-    #     ret[0] = pos_channel_buf[0] + pos_channel_buf[1]
-    #     mcp_pos[1] = 0.5 * MCP_CAL_Y * ( pos_channel_buf[2] - pos_channel_buf[3] ) * 0.001
-
-    
+    # comupte tofs in micro seconds 
     def compute_tof( self, current_trig_time, mcp_trigger_time ) :
         if mcp_trigger_time > current_trig_time : 
             return  ( mcp_trigger_time - current_trig_time ) * 25.0 / 1e6 
         else :
             return ( mcp_trigger_time + 2**24 - current_trig_time ) * 25.0 / 1e6 
+
         
+    # compute delay times in micro seconds
+    def compute_delay_times( self, delay_time_buf, mcp_trigger_time ) :
+        mask = ( mcp_trigger_time > delay_time_buf ).astype( int ) 
+        return ( delay_time_buf - mcp_trigger_time + mask * 2**24 ) * 25 / 1e6
 
-    def compute_timestamep( self ) :
-        pass
-
-
-    def apply_radius_cut( self ) :
-        pass
     
-    # apply cuts to the candidates found in process_candidates and store the results.
-    def apply_cuts( self ) :
-
-        # detect request for cut reset 
-        if self.num_cut_data == 0 :
-            start_idx = 0
-        else :
-            start_idx = self.num_events_prev 
-        
-        radii = self.radii[ start_idx : self.num_events ]
-        tofs = self.tofs[ start_idx : self.num_events ]
-
-        radius_mask = ( radii > self.r_cut_lower ) & ( radii < self.r_cut_upper )
-        tof_mask = ( tofs > self.tof_cut_lower ) & ( tofs < self.tof_cut_upper ) 
-
-        mask = radius_mask & tof_mask
-        indices = np.where( mask )[0] + start_idx
-        num_added_cut_data = len( indices ) 
-        
-        self.cut_data_indices[ self.num_cut_data
-                               : self.num_cut_data + num_added_cut_data ] = indices
-
-        self.num_cut_data += num_added_cut_data
-        
-
-    def reset_cuts( self ) :
-        self.num_cut_data = 0
-        self.num_events_prev = 0 
-        
-
-    def compute_polar( self ) :
-        
-        centered_mcp_positions = ( self.mcp_positions[ self.num_events_prev : self.num_events ]
-                                   - cpt_tools.mcp_center_coords  )
-        radii = np.linalg.norm( centered_mcp_positions, axis = 1 )
-        angles = np.degrees( np.arctan2( centered_mcp_positions[:,1],
-                                        centered_mcp_positions[:,0] ) )
-
-        self.radii[ self.num_events_prev : self.num_events ] = radii
-        self.angles[ self.num_events_prev : self.num_events ] = angles
 
         
     def save( self, path = None, name = None, prefix = None ) :
@@ -243,37 +231,72 @@ class LiveCPTdata( cpt_tools.CPTdata ) :
             path += '/data/' + prefix + '/' 
 
         if name is None :
-            name = create_data_name( self.tabor_params, prefix ) 
+            name = create_data_name( self.tabor_params, self.date_str, prefix ) 
 
         file_path = path + name
 
         print( 'INFO: saving file: ', file_path )
 
+        if os.path.exists( file_path ) :
+            cpt_tools.unlock_file( file_path )
+            os.remove( file_path ) 
+        
         try :
             self._save( file_path )
         except( OSError ) :
             os.makedirs( path, exist_ok = 1 ) 
             self._save( file_path )
             
-            
+
+
     def _save( self, file_path ) :
+        
+        # save_data = ( header_prefix, self.tabor_params.flatten(),
+        #               self.tofs[ : self.num_events ], self.timestamps[ : self.num_events ],
+        #               self.mcp_positions[ : self.num_events ], self.sums[ : self.num_events ],
+        #               self.sums[ : self.num_events ] )
 
+        tabor_params_labels = cpt_tools.TaborParams.flattened_header_labels() 
+        tabor_params_array = self.tabor_params.flatten() 
+
+        metadata_labels = [ 'Z', 'A', 'date', 'duration',
+                            'num_penning_ejects', 'num_mcp_hits', 'num_events' ]
+
+        metadata_array = [ self.Z, self.A, self.date_str, self.duration,
+                           self.num_penning_ejects, self.num_mcp_hits ]
+        
+        with open( file_path, 'w' ) as f :
+
+            # for i in range( len( metadata_labels ) ) :
+            #     line = '%s\t%f\n' % ( metadata_labels[i], metadata_array[i] )
+            #     f.write( line ) 
+
+            f.write( 'Z\t' + str( self.Z ) + '\n' )
+            f.write( 'A\t' + str( self.A ) + '\n' )
+            f.write( 'date/time\t' + self.date_str + '\n' )
+            f.write( 'duration (s)\t' + str( self.duration ) + '\n' )
+            f.write( 'num_penning_ejects\t' + str( self.num_penning_ejects ) + '\n' )
+            f.write( 'num_mcp_hits\t' + str( self.num_mcp_hits ) + '\n' )
+            
+            f.write( '\n' )
                 
-        #with open( file_path, 'wb' ) as f :
-        header_prefix = np.array( [ self.Z, self.A, self.timestamp, self.duration,
-                                    self.num_penning_ejects, self.num_mcp_hits ] )
-            
-        tmp = np.concatenate( ( header_prefix, self.tabor_params.flatten(), cpt_tools.cpt_header_key,
-                                self.all_data[ : self.num_events ].flatten() ) )
+            for i in range( len( tabor_params_labels ) ) :
+                line = '%s\t%f\n' % ( tabor_params_labels[i], tabor_params_array[i] )
+                f.write( line )
+                
+            f.write( '\n' )
+            f.write( 'x1\tx2\ty1\ty2\ttof\ttimestamp\tpenning_eject_index\n' ) 
 
-        try :
-            os.remove( file_path )
-        except OSError :
-            pass
+            save_array = np.vstack( [ self.delay_times[ : self.num_events ].T,
+                                      self.tofs[ : self.num_events ],
+                                      self.timestamps[ : self.num_events ],
+                                      self.penning_eject_indices[ : self.num_events ] ] ).T
             
-        tmp.tofile( file_path )
+            np.savetxt( f, save_array, delimiter = '\t', fmt = '%.8f\t%.8f\t%.8f\t%.8f\t%.4f\t%.4f\t%i' ) 
+                
         cpt_tools.lock_file( file_path ) 
-
+        
+                                  
             
 
 def create_element_prefix( Z, A ) :
@@ -293,12 +316,10 @@ def create_element_prefix( Z, A ) :
 
 
 # default data name 
-def create_data_name( tabor_params, prefix = None ) :
+def create_data_name( tabor_params, date_str, prefix = None ) :
 
     if prefix is None : 
         prefix = create_element_prefix( Z, A )
-
-    date_str = datetime.now().strftime( '%Y-%m-%d_%H-%M' ) 
         
     data_name = '%s_%s' % ( date_str, prefix )
